Crate wasmer_wasi_types

Wasmer’s WASI types implementation.

Those types aim at being used by the wasmer-wasi crate.

Modules

advice

directory

error

event

file

io

signal

snapshot0

subscription

time

versions

Structs

WasiSubscription

Safe Rust wrapper around __wasi_subscription_t

__wasi_ciovec_t

__wasi_dirent_t

__wasi_event_fd_readwrite_t

__wasi_event_t

__wasi_fdstat_t

__wasi_filestat_t

__wasi_iovec_t

__wasi_prestat_t

__wasi_prestat_u_dir_t

__wasi_subscription_clock_t

__wasi_subscription_fs_readwrite_t

__wasi_subscription_t

Enums

EventEnum

EventType

Safe Rust wrapper around __wasi_subscription_t::type_ and __wasi_subscription_t::u

PrestatEnum

SubscriptionEnum

Constants

__WASI_ADVICE_DONTNEED

__WASI_ADVICE_NOREUSE

__WASI_ADVICE_NORMAL

__WASI_ADVICE_RANDOM

__WASI_ADVICE_SEQUENTIAL

__WASI_ADVICE_WILLNEED

__WASI_CLOCK_MONOTONIC

__WASI_CLOCK_PROCESS_CPUTIME_ID

__WASI_CLOCK_REALTIME

__WASI_CLOCK_THREAD_CPUTIME_ID

__WASI_DIRCOOKIE_START

__WASI_E2BIG

__WASI_EACCES

__WASI_EADDRINUSE

__WASI_EADDRNOTAVAIL

__WASI_EAFNOSUPPORT

__WASI_EAGAIN

__WASI_EALREADY

__WASI_EBADF

__WASI_EBADMSG

__WASI_EBUSY

__WASI_ECANCELED

__WASI_ECHILD

__WASI_ECONNABORTED

__WASI_ECONNREFUSED

__WASI_ECONNRESET

__WASI_EDEADLK

__WASI_EDESTADDRREQ

__WASI_EDOM

__WASI_EDQUOT

__WASI_EEXIST

__WASI_EFAULT

__WASI_EFBIG

__WASI_EHOSTUNREACH

__WASI_EIDRM

__WASI_EILSEQ

__WASI_EINPROGRESS

__WASI_EINTR

__WASI_EINVAL

__WASI_EIO

__WASI_EISCONN

__WASI_EISDIR

__WASI_ELOOP

__WASI_EMFILE

__WASI_EMLINK

__WASI_EMSGSIZE

__WASI_EMULTIHOP

__WASI_ENAMETOOLONG

__WASI_ENETDOWN

__WASI_ENETRESET

__WASI_ENETUNREACH

__WASI_ENFILE

__WASI_ENOBUFS

__WASI_ENODEV

__WASI_ENOENT

__WASI_ENOEXEC

__WASI_ENOLCK

__WASI_ENOLINK

__WASI_ENOMEM

__WASI_ENOMSG

__WASI_ENOPROTOOPT

__WASI_ENOSPC

__WASI_ENOSYS

__WASI_ENOTCAPABLE

__WASI_ENOTCONN

__WASI_ENOTDIR

__WASI_ENOTEMPTY

__WASI_ENOTRECOVERABLE

__WASI_ENOTSOCK

__WASI_ENOTSUP

__WASI_ENOTTY

__WASI_ENXIO

__WASI_EOVERFLOW

__WASI_EOWNERDEAD

__WASI_EPERM

__WASI_EPIPE

__WASI_EPROTO

__WASI_EPROTONOSUPPORT

__WASI_EPROTOTYPE

__WASI_ERANGE

__WASI_EROFS

__WASI_ESPIPE

__WASI_ESRCH

__WASI_ESTALE

__WASI_ESUCCESS

__WASI_ETIMEDOUT

__WASI_ETXTBSY

__WASI_EVENTTYPE_CLOCK

__WASI_EVENTTYPE_FD_READ

__WASI_EVENTTYPE_FD_WRITE

__WASI_EVENT_FD_READWRITE_HANGUP

__WASI_EXDEV

__WASI_FDFLAG_APPEND

__WASI_FDFLAG_DSYNC

__WASI_FDFLAG_NONBLOCK

__WASI_FDFLAG_RSYNC

__WASI_FDFLAG_SYNC

__WASI_FILESTAT_SET_ATIM

__WASI_FILESTAT_SET_ATIM_NOW

__WASI_FILESTAT_SET_MTIM

__WASI_FILESTAT_SET_MTIM_NOW

__WASI_FILETYPE_BLOCK_DEVICE

__WASI_FILETYPE_CHARACTER_DEVICE

__WASI_FILETYPE_DIRECTORY

__WASI_FILETYPE_REGULAR_FILE

__WASI_FILETYPE_SOCKET_DGRAM

__WASI_FILETYPE_SOCKET_STREAM

__WASI_FILETYPE_SYMBOLIC_LINK

__WASI_FILETYPE_UNKNOWN

__WASI_LOOKUP_SYMLINK_FOLLOW

__WASI_O_CREAT

__WASI_O_DIRECTORY

__WASI_O_EXCL

__WASI_O_TRUNC

__WASI_PREOPENTYPE_DIR

__WASI_RIGHT_FD_ADVISE

__WASI_RIGHT_FD_ALLOCATE

__WASI_RIGHT_FD_DATASYNC

__WASI_RIGHT_FD_FDSTAT_SET_FLAGS

__WASI_RIGHT_FD_FILESTAT_GET

__WASI_RIGHT_FD_FILESTAT_SET_SIZE

__WASI_RIGHT_FD_FILESTAT_SET_TIMES

__WASI_RIGHT_FD_READ

__WASI_RIGHT_FD_READDIR

__WASI_RIGHT_FD_SEEK

__WASI_RIGHT_FD_SYNC

__WASI_RIGHT_FD_TELL

__WASI_RIGHT_FD_WRITE

__WASI_RIGHT_PATH_CREATE_DIRECTORY

__WASI_RIGHT_PATH_CREATE_FILE

__WASI_RIGHT_PATH_FILESTAT_GET

__WASI_RIGHT_PATH_FILESTAT_SET_SIZE

__WASI_RIGHT_PATH_FILESTAT_SET_TIMES

__WASI_RIGHT_PATH_LINK_SOURCE

__WASI_RIGHT_PATH_LINK_TARGET

__WASI_RIGHT_PATH_OPEN

__WASI_RIGHT_PATH_READLINK

__WASI_RIGHT_PATH_REMOVE_DIRECTORY

__WASI_RIGHT_PATH_RENAME_SOURCE

__WASI_RIGHT_PATH_RENAME_TARGET

__WASI_RIGHT_PATH_SYMLINK

__WASI_RIGHT_PATH_UNLINK_FILE

__WASI_RIGHT_POLL_FD_READWRITE

__WASI_RIGHT_SOCK_SHUTDOWN

__WASI_SHUT_RD

__WASI_SHUT_WR

__WASI_SIGABRT

__WASI_SIGALRM

__WASI_SIGBUS

__WASI_SIGCHLD

__WASI_SIGCONT

__WASI_SIGFPE

__WASI_SIGHUP

__WASI_SIGILL

__WASI_SIGINT

__WASI_SIGKILL

__WASI_SIGPIPE

__WASI_SIGPOLL

__WASI_SIGPROF

__WASI_SIGPWR

__WASI_SIGQUIT

__WASI_SIGSEGV

__WASI_SIGSTOP

__WASI_SIGSYS

__WASI_SIGTERM

__WASI_SIGTRAP

__WASI_SIGTSTP

__WASI_SIGTTIN

__WASI_SIGTTOU

__WASI_SIGURG

__WASI_SIGUSR1

__WASI_SIGUSR2

__WASI_SIGVTALRM

__WASI_SIGWINCH

__WASI_SIGXCPU

__WASI_SIGXFSZ

__WASI_SOCK_RECV_DATA_TRUNCATED

__WASI_SOCK_RECV_PEEK

__WASI_SOCK_RECV_WAITALL

__WASI_STDERR_FILENO

__WASI_STDIN_FILENO

__WASI_STDOUT_FILENO

__WASI_SUBSCRIPTION_CLOCK_ABSTIME

__WASI_WHENCE_CUR

__WASI_WHENCE_END

__WASI_WHENCE_SET

Functions

dirent_to_le_bytes

eventtype_to_str

print_right_set

function for debugging rights issues

right_to_string

expects a single right, returns None if out of bounds or > 1 bit set

wasi_filetype_to_name

Type Definitions

__wasi_advice_t

__wasi_clockid_t

__wasi_device_t

__wasi_dircookie_t

__wasi_errno_t

__wasi_eventrwflags_t

__wasi_eventtype_t

__wasi_exitcode_t

__wasi_fd_t

__wasi_fdflags_t

__wasi_filedelta_t

__wasi_filesize_t

__wasi_filetype_t

__wasi_fstflags_t

__wasi_inode_t

__wasi_linkcount_t

__wasi_lookupflags_t

__wasi_oflags_t

__wasi_preopentype_t

__wasi_riflags_t

__wasi_rights_t

__wasi_roflags_t

__wasi_sdflags_t

__wasi_siflags_t

__wasi_signal_t

__wasi_subclockflags_t

__wasi_timestamp_t

__wasi_userdata_t

__wasi_whence_t

Unions

__wasi_event_u

__wasi_prestat_u

__wasi_subscription_u